Refactor, move config to model package.
